Move GroupBy to separate module
